requirements.txt
anyjson==0.3.3
beautifulsoup4==4.5.1
cffi==1.11.5
crypto==1.4.1
cryptography==1.4
decorator==4.0.10
Django==1.11.2
django-extensions==1.7.2
django-model-utils==2.5.2
django-notifications-hq==1.2
django-ratelimit==1.1
httplib2==0.9.2
idna==2.6
jsonfield==1.0.3
Naked==0.1.31
numpy==1.11.1
Pillow==3.3.0
psycopg2==2.6.2
pyasn1==0.4.2
pycparser==2.18
pycrypto==2.6.1
pytz==2018.3
PyYAML==3.12
ratelimit==1.4.1
requests==2.10.0
shellescape==3.4.1
six==1.11.0
sorl-thumbnail==12.3
